VSM-39 + VSM-47
VSM-39 | Adding cards, continued
Added logic for adding main-activities. Still waiting on an API-change to correct the sorting.
Also, insert is waiting on api-response.
Maybe we should improve this to feel more snappy for the user...?
Or show a spinner or something while waiting. (even though its's only a couple of seconds on my line)
https://equinor-sds-si.atlassian.net/browse/VSM-39

VSM-47 | Delete vsmObjects
https://equinor-sds-si.atlassian.net/browse/VSM-47

+ Refactor local vsmObject patch-operation `patchNode()`
+ Rename vsmObject