Implement new VSM structure* + change dsiplay->display key (#32)
Implement new VSM Structure
https://equinor-sds-si.atlassian.net/browse/VSM-74

Initial implementation of displaying choice-children. Still pretty easy to mess up the layout. Let’s check if this is good enough for MVP. 
https://equinor-sds-si.atlassian.net/browse/VSM-42

Deleting cards should not delete the children
> The new structure makes that this is working as expected.
https://equinor-sds-si.atlassian.net/browse/VSM-51

Typo in displayKey is now updated in the api, so we fix it in app as well.
https://equinor-sds-si.atlassian.net/browse/VSM-67?focusedCommentId=25971