Role should be only one line
THOKN@equinor.com reported a bug with ID 71:In sub-activities; when the role is above a certain length it is divided over two lines. The second line interrupts with the time input on the cards. An example from project 139 is the role “Operation engineer”. The word “Operation” sits in the first row, while the second row has “engine” in it, on top of the time.

https://equinor-sds-si.atlassian.net/browse/VSM-163