feat: to-be process
**Background**
A `to-be` process is an alternative of the same `original` process, that is mapped how someone imagines a _**perfect**_ ✨ process to be.
The user should be able to quickly change between these two alternatives and compare them together.
The `to-be` process is technically and practically a new process with its own Problems, Questions, Ideas and Risks. Though it shares the same name, labels and access-roles.

**Requirements**
- [x] User can flip a toggle to switch between the `original` and the `to-be` process.
- [x] Name is shared
- [x] Access is the same as the original process
- [x] ~Labels are shared between the `original` and the `to-be` processes.~ Labels are not yet implemented in the front-end, but should work... #159
- [x] To-be processes should not be shown on the front-page. _(We need some metadata to distinguish them from regular ones...)_
- [x] Individual version history.
- [x] PQIRs are not shared between `original` and `to-be`.
- [x] When duplicating a process, it should always duplicate the `current`-process. (Not the `To-be`-process)

**Use-case**
The API handles all requests related to labels, process-name, and access to the `to-be` process it as if it was the `original` process.

Ex 1.
Client ask for labels for a `to-be` process.
The API answers with the labels for the `original`.

Ex 2
Client adds a label to a `to-be` process.
The API handles it as if it was the `original` process.

closes #192