chore: Use RADIX_GIT_COMMIT_HASH environment variable instead of hack (#381)
Adjusting the tagging so that the RADIX_GIT_COMMIT_HASH is correct.
This simplifies our process. Thus removing the changes I intruduced in #370

I think it was making it an annotated tag that made it not work.  
Using Lightweight Tags seems to be working as expected. :slightly_smiling_face: 
> To create a lightweight tag, don’t supply any of the -a, -s, or -m options
https://git-scm.com/book/en/v2/Git-Basics-Tagging

fixes #275